+++ server.R
@@ -1,7 +1,7 @@
 
 pacman::p_load(shiny, shinyjs, shinyWidgets, magrittr, dplyr, datimvalidation, ggplot2,
                futile.logger, paws, datapackr, scales,
-                DT, purrr, praise, rpivotTable, waiter, flextable, officer, gdtools)
+                DT, purrr, praise, rpivotTable, waiter, flextable, officer, gdtools,digest)
 
 
 #Parallel execution of validation rules on Windows is not supported
@@ -75,8 +75,6 @@ shinyServer(function(input, output, session) {
     r  <-  saveTimeStampLogToS3(d)
     timestampUploadUI(r)
     archiveDataPackErrorUI(r)
-    r  <-  sendValidationSummary(d, "validation_error")
-    validationSummaryUI(r)
     r  <-  saveDATIMExportToS3(d)
     waiter_hide()
     datimExportUI(r)
@@ -139,7 +137,7 @@ shinyServer(function(input, output, session) {
   observeEvent(input$kpCascadeInput, {
     kpCascadeInput_filter$snu_filter  <-  input$kpCascadeInput
   })
-
+  
   observeEvent(input$logout, {
     flog.info(paste0("User ", user_input$d2_session$me$userCredentials$username, " logged out."))
     ready$ok  <-  FALSE
@@ -281,6 +279,7 @@ shinyServer(function(input, output, session) {
                    h5("Note: This is a draft memo table. Final figures may differ."),
                    tags$h4("Data source: PSNUxIM tab")),
           tabPanel("Memo Comparison",
+                   fluidRow(actionButton('reset_pivot',"Reset pivot")),
                    fluidRow(column(width = 12,
                                    div(rpivotTable::rpivotTableOutput({"memo_compare"})))), # nolint
                    fluidRow(tags$h4("Data source: PSNUxIM tab & DATIM")))
@@ -352,7 +351,10 @@ shinyServer(function(input, output, session) {
 
   output$memo_compare  <-  renderRpivotTable({
     vr <- validation_results()
-
+    #Take a dependency on the reset button
+    
+    reset_pivot <- input$reset_pivot
+    
     if (!inherits(vr, "error") & !is.null(vr)) {
 
       if (is.null(vr$data$compare)) {
@@ -603,17 +605,18 @@ shinyServer(function(input, output, session) {
       d <- validation_results()
 
       if (input$downloadType  == "messages") {
+        sendEventToS3(d,"MESSAGE_DOWNLOAD")
         writeLines(d$info$messages$message, file)
       }
 
       if (input$downloadType  == "cso_flatpack") {
-
+        sendEventToS3(d,"CSO_FLATPACK_DOWNLOAD")
         wb <- downloadCSOFlatPack(d)
         openxlsx::saveWorkbook(wb, file = file, overwrite = TRUE)
       }
 
       if (input$downloadType  == "flatpack") {
-
+        sendEventToS3(d,"FLATPACK_DOWNLOAD")
         waiter_show(html = waiting_screen_flatpack, color = "rgba(128, 128, 128, .8)")
         datapack_name  <- d$info$datapack_name
         flog.info(
@@ -632,6 +635,7 @@ shinyServer(function(input, output, session) {
         sheets_with_data <- d$tests[lapply(d$tests, NROW) > 0]
 
         if (length(sheets_with_data) > 0) {
+          sendEventToS3(d,"VR_RULES_DOWNLOAD")
           openxlsx::write.xlsx(sheets_with_data, file = file)
         } else {
           showModal(modalDialog(
@@ -651,6 +655,7 @@ shinyServer(function(input, output, session) {
         flog.info("Fetching support files")
         d <- downloadDataPack(d)
         openxlsx::saveWorkbook(wb = d$tool$wb, file = file, overwrite = TRUE)
+        sendEventToS3(d,"DATAPACK_DOWNLOAD")
         flog.info(
           paste0("Datapack reloaded for for ", d$info$datapack_name),
           name = "datapack")
@@ -672,11 +677,12 @@ shinyServer(function(input, output, session) {
                             sheet = "Comparison", x = d$data$compare)
 
         openxlsx::saveWorkbook(wb, file = file, overwrite = TRUE)
+        sendEventToS3(d,"COMPARISON_DOWNLOAD")
         waiter_hide()
       }
 
       if (input$downloadType  == "memo") {
-        
+        sendEventToS3(d,"MEMO_DOWNLOAD")
         doc <- downloadMemo(d)
         print(doc, target = file)
 
@@ -685,6 +691,7 @@ shinyServer(function(input, output, session) {
   )
 
   validate <- function() {
+    
     shinyjs::disable("downloadType")
     shinyjs::disable("downloadOutputs")
     shinyjs::disable("send_paw")
@@ -730,10 +737,10 @@ shinyServer(function(input, output, session) {
         d$info$approval_status <- "UNAPPROVED"
         #Generate a unique identifier
         d$info$uuid <- uuid::UUIDgenerate()
+        #Get a single operating unit from the country IDs
         d$info$operating_unit <- getOperatingUnitFromCountryUIDs(d$info$country_uids)
-        
-        
-        
+        #Log the validation to S3
+        sendEventToS3(d,"VALIDATE")
         flog.info(paste0("Initiating validation of ", d$info$datapack_name, " DataPack."), name = "datapack")
         if (d$info$tool  == "Data Pack") {
           
@@ -744,6 +751,7 @@ shinyServer(function(input, output, session) {
                             choices = downloadTypes(tool_type =  d$info$tool,
                                                     needs_psnuxim = d$info$needs_psnuxim,
                                                     memo_authorized = user_input$memo_authorized))
+          
           if (d$info$has_psnuxim & NROW(d$data$SNUxIM) > 0) {
             
             flog.info(paste(d$info$tool, " with PSNUxIM tab found."))
@@ -786,7 +794,8 @@ shinyServer(function(input, output, session) {
             d <- comparePrioTables(d)
             Sys.sleep(1)
             incProgress(0.1, detail = ("Finishing up."))
-            r <- sendValidationSummary(d, "app_analytics", include_timestamp = TRUE)
+            flog.info("Sending validation summary")
+            r <- sendValidationSummary(d, "validation_error_summary", include_timestamp = TRUE)
             validationSummaryUI(r)
             
             shinyjs::enable("downloadType")